moving user init to connecting wallet
